requirements.txt
langchain 
langchain-community 
langchain-openai 
langchain-experimental 
langchain-neo4j
neo4j 
wikipedia 
tiktoken 
yfiles_jupyter_graphs
python-dotenv
json-repair